separate triton python module into multiple files
